+++ spring-telqos-core/src/main/java/com/dwarfeng/springtelqos/node/configuration/SpringTelqosHandlerDefinitionParser.java
@@ -23,13 +23,13 @@ public class SpringTelqosHandlerDefinitionParser implements BeanDefinitionParser
 
     @Override
     public BeanDefinition parse(Element element, @Nonnull ParserContext parserContext) {
-        String handlerName = BeanDefinitionParserUtil.mayResolvePlaceholder(
+        String handlerName = (String) BeanDefinitionParserUtil.mayResolveSpel(
                 parserContext, element.getAttribute("handler-name")
         );
-        String executorRef = BeanDefinitionParserUtil.mayResolvePlaceholder(
+        String executorRef = (String) BeanDefinitionParserUtil.mayResolveSpel(
                 parserContext, element.getAttribute("executor-ref")
         );
-        String configRef = BeanDefinitionParserUtil.mayResolvePlaceholder(
+        String configRef = (String) BeanDefinitionParserUtil.mayResolveSpel(
                 parserContext, element.getAttribute("config-ref")
         );