+++ spring-telqos-core/src/main/java/com/dwarfeng/springtelqos/sdk/util/BeanDefinitionParserUtil.java
@@ -1,6 +1,11 @@
 package com.dwarfeng.springtelqos.sdk.util;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.xml.ParserContext;
+import org.springframework.expression.Expression;
+import org.springframework.expression.ExpressionParser;
+import org.springframework.expression.common.TemplateParserContext;
+import org.springframework.expression.spel.standard.SpelExpressionParser;
 
 /**
  * Spring XML 命名空间解析器工具类。
@@ -19,6 +24,46 @@ public final class BeanDefinitionParserUtil {
      */
     public static String mayResolvePlaceholder(ParserContext parserContext, String attribute) {
         return parserContext.getReaderContext().getEnvironment().resolvePlaceholders(attribute);
+    }
+
+    /**
+     * 如果指定的属性是一个 SpEL 表达式，则解析它，否则返回原属性。
+     *
+     * <p>
+     * 需要注意的是，由于 SpEL 解析后的结果可能是一个对象，因此该方法的返回值类型为 Object。
+     * 对象的类型需要进行约定，例如约定 SpEL 表达式解析后的结果必须是一个字符串，
+     * 或者约定 SpEL 表达式解析后的结果必须是一个特定类型的对象。
+     *
+     * <p>
+     * 需要注意的是，使用本项目解析 SpEL 表达式时，Spring 的状态仍处于初始化阶段，
+     * 因此无法使用 Spring 的 BeanFactory 来解析 SpEL 表达式。<br>
+     * 本方法是在 Spring 的解析机制之外，额外创建了一个 SpEL 解析器来解析 SpEL 表达式的。<br>
+     * 因此，使用本方法解析 SpEL 表达式时，无法使用 Spring 的 BeanFactory 来解析 SpEL 表达式中的占位符，
+     * 如无法使用 Spring 的 BeanFactory 提供的 {@link org.springframework.expression.ParserContext}。<br>
+     * 因此，在任何可能的情况下，建议先使用 {@link #mayResolvePlaceholder(ParserContext, String)} 方法解析占位符。
+     * 比如：
+     * <blockquote><pre>
+     * BeanDefinition.getPropertyValues().add(propertyName, propertyValue)
+     * </pre></blockquote>
+     * 其中，<code>propertyValue</code> 既可以直接使用 SpEL 表达式本身，也可以解析 SpEL 表达式后直接使用解析结果。
+     * 此时更推荐直接使用 SpEL 表达式本身。
+     *
+     * @param parserContext Parser 上下文。
+     * @param attribute     指定的属性。
+     * @return 也许被解析的属性。
+     */
+    public static Object mayResolveSpel(ParserContext parserContext, String attribute) {
+        String resolvedAttribute = parserContext.getReaderContext().getEnvironment().resolvePlaceholders(attribute);
+
+        String trimmed = StringUtils.trim(resolvedAttribute);
+        if (StringUtils.isEmpty(trimmed) || !trimmed.contains("#{") || !trimmed.contains("}")) {
+            // 如果解析后的字符串不包含 SpEL 表达式的标志，则直接返回解析后的字符串。
+            return resolvedAttribute;
+        }
+
+        ExpressionParser parser = new SpelExpressionParser();
+        Expression expression = parser.parseExpression(resolvedAttribute, new TemplateParserContext());
+        return expression.getValue();
     }
 
     /**